requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-Levenshtein>=0.21.0
idna>=3.4
requests>=2.28.0
python-telegram-bot>=20.0
pyzbar>=0.1.9
Pillow>=9.0.0
python-dotenv>=1.0.0
